Track cross-TextBuffer changes to check TextIter validity
In our async callbacks for updating inline highlighting, we have no
guarantee that the TextIters passed in are still valid. One option
would be to use TextMarks instead, but this isn't actually what we want;
we actually want the TextIters to invalidate, and not highlight when
they're not valid.

Since GTK+ doesn't expose a check for the validity of TextIters, we
reproduce a dodgy version here. We keep a counter of changes to all
TextBuffers, and check that the counters match in the highlighting
callback.