Add recently-used comparison support (closes bgo#652747)
The recent-files API provided by GTK+ doesn't actually work for Meld
out-of-the-box, because instead of storing individual files, we need to
stored multiple linked files. In other words, we need to store a
comparison, not a file.

This commit adds support for reading and writing a simple comparison
record format with a new Meld-specific mime-type. These files are
stored under the user data directory, and are managed by the new
'recent' module. Meld creates these files for new top-level
comparisons and inserts them into the recently-used store as proxies
for the actual file tuples.

Note that we deliberately avoid recording as recently-used comparisons
that are invoked from other comparisons; a user may open ten quick
file comparisons from a single VC comparison, but they probably don't
actually want to re-open those from the recent files menu.

There is also support for opening comparison files from the command
line. This was added so that recent comparisons can be opened from the
desktop recent files menu, but can also be used to manually open
specified comparisons.