Adjust VC location semantics to require a location to be a folder
The VC location field was only ever used to set cwd for various
operations, to limit comparisons to subsections of a repository.
While in theory we could use it to identify single-file comparisons,
this was never done and isn't obviously useful.

This commit changes the semantics such that the location field is
now the folder for which the comparison was launched, regardless of
whether it is a sub-folder or a single-file comparison.