Fix VC listing when we have an empty directory
Previously, if we had no files or folders in a directory, we skipped
a lookup step and simply returned nothing. This breaks when we have
nothing but MISSING or REMOVED files in the directory, which should be
shown but are instead simply skipped. This commit changes our scanning
to always run a lookup even for empty directories.