Adjust _vc.call() to explicitly discard all subprocess output
Since call() should only ever be used to check the return value of a
command, there's no point asking for a pipe to the output. In fact,
asking for a pipe causes issues when the output exceeds the internal
buffers and stalls. This patch remedies both the semantic and practical
problem by explicitly dropping stdout and stderr into os.devnull.