Support setting TextViews to read-only (closes bgo#321637)
While we already have support for handling read-only buffers,
introduced as part of auto-merge mode, we haven't actually used this in
regular comparisons until now. The reason it wasn't used is that
editing files you can't write to can still be useful in resolving
conflicts, by eliminating or simplifying differences.

In order to accommodate both the use case of simpler and more
consistent actions, and the occasional need to edit read-only files,
this commit introduces a toggleable lock indicator for read-only files
which both indicates read-only status (and thus obsoletes our previous
read-only indicator) and allows the user to lock or unlock read-only
files.

This commit also changes default behaviour, so that read-only files are
now uneditable by default.