Handle recent-file reading of malformed files (closes bgo#690472)
While it's not entirely clear how this could happen, some non-meldcmp
files can end up in Meld's recent file list. While we check for
configparser files that don't match the layout we expect, we don't
check for recent files that aren't even configparser files. This commit
adds such a check.

It's entirely possible that the cause of this problem is simply
development changes, but the error handling is a good idea anyway.