Use cursor position change rather than manual trigger for diff recalc
Manually triggering a diff-change when a buffer was set editable was
problematic in cases where we're loading a new file. In this situation,
the current diff chunk was set to an invalid index in the newly-reset
differ, as the current chunk isn't invalidated.

This commit forces the diff-change recalculation by instead forcing a
cursor chunk change, which is how we do this elsewhere in the
codebase.