Fix overzealous inline highlighting invalidation (closes bgo#693690)
In the existing code, any buffer change invalidated all pending inline
highlighting jobs. This causes serious issues if any highlighting jobs
are pending when a change is made; these jobs will never be restarted
unless their chunks are directly changed.

The solution in this patch is to only consider inline highlighting
results to be invalid if the actual text being highlighted has changed.
We do this by using gtk.TextMarks instead of gtk.TextIters (avoiding
buffer invalidation) and checking to see whether the text we're about
to highlight actually matches what we think we're highlighting. If it
doesn't, then we have a changed block, which *will* be rehighlighted
in the relevant changed callback.

The downside to this approach is that creation of gtk.TextMarks and
validating the text before highlighting is computationally much more
expensive than our previous approach.