Add is_temp return to 'get_path_for_conflict' for non-temporary files
Depending on the VC, we sometimes create new temporary files using
direct VC calls, and sometimes reuse existing files from the conflict.
This patch adds an boolean return to get_path_from_conflict to
indicate whether the returned file is temporary and should be cleaned
up, or whether we should leave it alone.