Fix adding directories on CVS and SVN (closes bgo#501098)
Adding directories in CVS and SVN doesn't work because these need to be
added from the parent folder. This patch adds a hack to VcView to
adjust our behaviour specifically for CVS and SVN modules to allow
directories to be added separately, and from a custom parent path.

Based on testing, Subversion 1.7 appears to handle this case without
special treatment, so this fix is applied only to the old SVN module.

Based on a patch from Manual Siggen.