Apply blank line ignoring for folder comparisons (closes bgo#695912)
Previously, we applied any active filters, but didn't include the
ignore-blank-lines preference. As a result, the hack of filtering and
then ignoring blank lines as a substitute for multi-line regexes didn't
work as expected.

This patch changes the folder comparison to filter blank lines after
filtering is done, if so configured.