Fix some handling of unicode file names (closes bgo#168760, bgo#694099)
While in theory we should be using encoding from the file system, this
turns out to usually be incorrect. Typically, the filenames we handle
have already passed through GTK+ (and are thus UTF-8) or have come
from Python, in which case if we pass unicode in, we get unicode out.

This patch basically takes these two factors into account and tries
to make sure we give Python's file system handling unicode objects and
decodes the UTF-8 we get from GTK+. This appears to work on both
Linux (which it already did, assuming a UTF-8 file system...) and on
Windows (UTF-16, probably).