Fix universal newline support on Windows (closes bgo#687729)
The codecs module uses and/or doesn't use Python's universal newline
support in odd ways, and in particular breaks our assumptions on
Windows. This patch changes from using the codecs module to using the
newer io module to do the same job, with the bonus that this should be
more forwards-compatible.