Make Git 'update' command (pull) path-less (bgo#698564)
Since it's not actually possible to do this under the existing VC
interaction model, this commit includes the start of a new interface
to the VC modules. In this interface, the command is called by the
VC module, which is passed a 'runner' - essentially just a method
that takes a command list and a files list and runs the result.

The runner interface will probably have to change to require a
subprocess-style list of all arguments, but the current way is much
easier to hook in to existing code.