Remove Reload, replacing with Revert for file comparisons (bgo#699048)
For version control and folder comparisons, Reload and Refresh already
did exactly the same thing. For file comparisons, Reload actually
re-reads files off disk, discarding changes if there were any. Since
this behaviour is closer to that of gedit's Revert, this commit
removes the Reload action for all views, and adds a Revert command to
the File menu for file comparisons.