Add unpushed commits summary for Git (bgo#699023)
This change adds a method to the new VC API that allows a VC to
populate its root folder with some metadata, currently supported as
the unpushed commits in a repository. Currently only Git uses this
capability; hg outgoing would do the same very nicely, but since it
moves data over the network, this isn't really acceptable.