Improve sensitivity handling by delegating to VC modules (bgo#698511)
Currently all of our sensitivity handling in VC views is done purely
on the basis of the capabilities of a VC, and whether or not the view
has a selection. This gives bad information in many cases, such as
allowing an unversioned file to be reverted.

This commit adds new API to the VC module interface that delegates
the decisions about what VC actions to enable for a given selection.
Currently only Git support this interface, but it's designed to be
very VC-neutral.