Don't uncache the VC inventory
The memory savings here are very very minor anyway, and it was almost
always a loss, as we often then re-cache soon thereafter. Most
importantly, this exposed a bug in our SVN usage (or rather in the API
itself) where the first re-cache was done with a subdirectory and we
had a bad cache for the rest of the tree.