Fix change categorisation around synchronisation points (bgo#699399)
Previously we did our synchronisation point split, but when handing off
the result to diffutil we removed the split. This resulted in weird
situations where what should have been paired inserts were merged into
a single changed block.

This change makes it so that we keep an internal copy of our split
results and use that when getting the opcodes.

This also half-fixes bgo#699067 by changing to use gtk.TextMarks for
sync points, but editing still breaks things.