Add preference to swap VC pane order (closes bgo#379160, bgo#699030)
Maybe it would have been better if Meld always used the left=local
model, but given that it's been the opposite way forever, we can't just
swap panes around on people.

This commit adds a preference for whether left or right is local,
defaulting to the current behaviour. This also applies to conflict
comparisons.