vc: Move repository root finding to class and adjust subclasses
The way Meld has previously dealt with figuring out whether a path is
a valid repository is to try to create the Vc, and then catch the
resulting explosion if it fails. While this works, it ends up having
some very weird side effects in VcView, and makes for some overly
complicated logic when figuring out what VC to use.