Change VC root selection for VC's that don't walk to find the root
When invoked on subdirectories of a repository with both root-walking
and non-root-walking VCs, we get weird situations. For example, in
a CVS + GIT repo rooted at foo/, every folder under foo will have a
CVS folder, but only foo/ will have a .git folder. This means that
our only-get-the-most-specific-VCS heuristic is horribly broken when
invoked on foo/wibble/.

The 'fix' here is to simply not consider the depth of VCs that don't
root walk when deciding whether to clear our most-specific-VC list.
This is still broken for subversion because 1.6 doesn't root walk,
but 1.7 does, and they both look for the same directory. However,
it should ameliorate the problem for CVS users.