filediff: Don't disable updating when syncpoints aren't yet active
This is a problem if, for example, the user sets a syncpoint in the
left pane without setting one in the right pane. We need matching
syncpoints in each pane to do the syncing so don't actually sync and
don't show our warning, but we do disable comparison updating.

This patch changes this so that we only disable comparison updating
if the actual differ has a list of syncpoints that it's using.