diffmap: 'Fix' Diffmap sizing for some weird DirDiff race condition
This isn't a real fix; for some reason even when a resize is queued
with the correct width, the draw event comes through while there is
still an old requisition on the widget. This isn't reproducible if
the allocation is >1, so it looks like it might be some weird
optimisation edge case?