meldapp: Integrate with OptionParser and output to local stdout
This fixes some issues with new tab/window creation and lifecycles,
as well as introducing a godawful hack to integrate GApplication
and GApplicationCommandLine with optparse. Without support for local
command line handling this is never going to be pretty.

There are several remaining slightly weird issues here, but it works
and most people hopefully won't hit the edge cases. Better support
is coming in glib 2.40 for doing this sort of thing, so let's hope that
pans out.