filediff: Fix interaction between set labels and save-as action
When setting fixed labels, we don't expect the user to change the
file, and there's no UI for doing so. However, it's still possible
(and reasonable) to change the file path by saving a labelled pane as
a different path. In this case, we already correctly set the path on
the file selector, so we just need to re-show it, in case the pane is
actually labelled.