vcview: Support setting tab label, improve version labels (bgo#699056)
Having good labels that work for all VCs is hard, so these strings are
a compromise. Setting the tab label automatically based on the strings
used for the pane labels isn't sensible, so we now specify that as
well.

This commit also fixes some very confused ordering of local/remote
labels for versions.