diffgrid: Allow resizing file and folder comparison panes (bgo#576934)
To individual adjust the width of each pane in a file or folder
comparison is something that can be pratical good to have. But by using
GtkGrid for layout its view creates a limitation of not allowing pane
resizing. Here is an implementation of a GtkGrid based custom widget,
by overriding its childrens size allocation and adding in the row at
the top of each LinkMap a drag handle, the size and position can easy
be set.