build_helpers: Override install command to fix install layout issues
Debian and Ubuntu need --install-layout=deb to be enabled by default
for sane user behaviour, but that flag doesn't work for non-install
commands, breaking `python setup.py build`.

This commit stops hard-adding the layout argument to sys.argv for
Debian and Ubuntu in favour of defaulting the argument to the deb
layout if it's not otherwise specified, and only doing so on install.