meldapp: Only apply the auto merge flag to the last comparison started
When more than one comparison is started with --diff, the --auto-merge
flag should only apply to the final one, which will always be the
one given with normal args, if such exists. Any way we do this is
slightly odd, but this fix preserves the correct behaviour for a fairly
common merge helper invocation with a single auto-merged three way and
two two-way diffs to ancestors. It's also closer to what we did in
1.8.x than applying auto-merge to everything.