Make GtkSourceView scheme-style configurable
This commit adds a "color scheme" entry to preferences which controls
the GtkSourceView scheme-style property to allow the user toggle between
different color schemes installed in the system. This is particularly
relevant for syntax highlighting since the default scheme (classic) is
not great in all circumstances such as dark theme usage. In general this
is heavily user preference driven so this commit allows users to select a
scheme they like.