bin/meld: Add hacky CSS for GtkTextView drawing changes in 3.13+
As on GTK+ 3.13.something, GtkTextView now provides us with a
draw_layer vfunc that we can use instead of just drawing on our
SourceView in the callback and assuming that it will work. However, as
part of this change, the background is now also reset when drawing,
which has broken our existing drawing mechanism (see bgo#737957).

Until we can adapt to the new drawing model (and require GTK+ 3.14 for
it) this hack just makes the background of our SourceView transparent,
so that our painting on the lower layer is still shown.