filediff: Fix bug with very many opcodes being cycled through the cache
In some situations where we, for example, did a replace all that caused
many inline highlighting requests to pile through the cache, it was
possible that between checking that we had the cached opcodes and
actually returning those opcodes in the mainloop, the opcodes would be
evicted from the cache. With this change, we keep an in-scope copy of
the cached opcodes around so that this won't happen.