bin/meld: Work-around for GTK+ 3.14 background drawing, part 2
Commit 61f53b0c reset the drawn background to transparent, which works
for most drawing, but looks terrible as soon as any text is selected.

This commit adds a secondary horrible hack that tries to reproduce the
selected text background colour once we've reset the main widget
background. This does use a horrible undocumented internal CSS class
that GtkTextView applies when rendering text, but it's the only way
that I could find that should hopefully work across different themes.