meld.vc.git: Fix path for modified files from a subdir (bgo#738561)
In _update_tree_state_cache(), we now go to some effort to get an
absolute, unicode file path, assuming that we have paths relative to
our location (i.e., not the repo root). However, our calls to git to
get paths for modified and staged files were only limiting the returned
files to the location that we're scanning, *not* returning paths
relative to that location.

This change simply adds the --relative flag before our existing path
argument, which fixes the issue.