sourceview: Fix GtkSourceView tab width property binding (bgo#738001)
The tab-width and indent-width properties interact weirdly if they're
different. This was a breakage in the GtkSourceView 3 migration, so
just switch the binding back over here.

This might cause some pain in the future if we want to support setting
both for some reason, but I don't actually know why we'd want to do
that.