bin/meld: Recompile gschemas based on schema mtime check (bgo#739267)
When checking whether we have a compiled version of the schemas, we now
also check to see whether it looks like it's actually current, based on
the relative mtimes of the source and compiled schema files.

This keeps biting people running from source whenever we change the
schemas. While this fix isn't exactly elegant, it's better than
constantly running into the problem, since that's what the
autocompiling-the-schemas hack was added for in the first place.