sourceview: Update drawing to use cairo clip extents for bounds
Previously, we checked to see what portion of the TextView was visible
when deciding which changes needed redrawing. That doesn't work now
with the pixel cache, which draws both above and below what's actually
on screen.

With this change, we trust that the cairo clip extents match up with
what's going in to the pixel cache, and so use those as bounds for our
change drawing.