sourceview: Change binding time to fix bad gutter drawing (bgo#739927)
Why, my future self may ask, does this exist? What the hell? Well,
see... here's how it is.

In the past, we set all of these properties on a GtkSourceView
manually, and GtkSourceView lazily instantiated its gutter objects and
connected their signal handlers and everything was fine. Then we
started doing construct-time binding with gsettings, which changed when
the gutter was created. This then meant that instead of our signal
handler for the textview's 'draw' signal being connected *before* the
signal handler for doing gutter drawing it was connected *after*, but
*only when* the show-line-numbers preference was active when we created
our MeldSourceView object. The result was that our signal handler drew
all over the GtkSourceGutter drawings and everyone was sad.

The 'solution' is to *not* bind show-line-numbers during GObject construction
and instead bind it after we've connected our other signal handler.