filediff: Skip unnecessary textview drawing and remove vestigal code
This is only a partial fix for unnecessary drawing, but should skip a
small amount of pointless work. Doing this better isn't actually worth
it because the pixel cache and draw-layer changes in GTK+ 3.14 just
make this so much faster it's not even funny.

Also, the hack in the commented-out code isn't necessary any more, so
remove it. How did it even escape detection?!