filediff: Repaint TextView background for modified CSS (bgo#741287)
In cases when we've messed with our TextView's CSS to fix GTK+ 3.14
regressions, we need to also repaint the background because some themes
don't otherwise do so. Specifically, this breaks oxygen-gtk3 such that
diff regions have a background colour, but the rest of the text view
is unpainted, which is pretty terrible.