filediff: Add change deletion pushing, reducing pane swaps (bgo#729838)
Previously we enforced the action of pushing a change to mean that
there had to be some text to push. As it's been requested a few times,
this patch makes it so that it's possible to push an empty change as
well, which has the effect of deleting the change from the target pane.

This feels slightly odd mental-model-wise, but pragmatism beats purity
and this means that you have to switch panes much less frequently.