filediff: Reset chooser filename for cancelled selection (bgo#741928)
If the user has an unsaved file and selects a new file in the chooser,
we prompt for saving the existing file. If that dialog is cancelled,
we're left in an inconsistent state where the file chooser thinks
we've got one file, and the buffer has another.

This commit adds a slightly bodgy workaround for this situation, just
manually checking whether the action was cancelled. It would be
possible to do something more serious by performing our checks before
the chooser dialog closes, but that would be a lot more work for little
gain.