meldapp: Handle opening ambiguous relative paths (bgo#742142)
Because GApplicationCommandLine.create_file_for_arg() handles absolute
and relative paths *and* URIs, it will try to interpret some relative
path names (e.g., foo:bar) as a URI.

This patch fixes this by checking to see whether the created file
exists, and if not attempts to treat it as a relative path name. The
extra cwd handling is necessary due to GApplication remote invocation.