dirdiff: Expand the root path regardless of whether there were changes
It's always looked a bit weird that we didn't expand the root path,
because it's not clear that we've finished and just haven't... you
know... not expanded anything. With this change, we expand the root
path when we finish so that the user can more easily see that we've
actually looked. Together with the new notification, this should make
the no-differences case a lot nicer.