meldbuffer: Fix splitlines hack for multiple line ends (bgo#743301)
The hack this fixes exists because Python's splitlines() splits on a
few characters that GTK+'s text handling doesn't. The hack 'correctly'
handles this case as long as there's only one of these characters in
a line, but in files with e.g., multiple 0x0c characters on a single
line, it broke.

This change fixes the problem by not moving off a line until we go
through without performing any of our 'fixes' to it.