dirdiff: Rudimentary support for blank folder comparisons (bgo#742342)
As a consequence, we also now don't compare a user's home directory to
itself if the user starts a comparison without selecting any folders.

This 'support' still requires there to be a blank, non-existant entry
in the comparison tree, but otherwise basically works as expected.
Getting rid of that blank entry would be nice, but basically everywhere
in the dirdiff/tree model stack expects to get at least one valid path
per tree row, so it's a bunch of churn for minimal gain.