filediff: Handle save responses better in delete events (bgo#744081)
The "you have modified files" dialog offers a "Close without saving"
option, which returns a CLOSE response code instead of our usual OK.
This was being handled correctly in most places, but not when the tab
was being closed via a delete event. This patch fixes it so that the
check_save_modified() always returns either OK or CANCEL, since that's
really the only thing we need out of it.

This commit also removes a stray delete-event handler on a widget that
should never ever be loaded.