preferences: New pref for applying text filters in folder comparison
New preferences are awful, but this has been a long-standing issue for
performance of folder comparisons, and it's often not clear to new
users that it's a thing we do. Also, the upcoming newline-ignoring
change makes this even more expensive, so now seems like the right time
to add this.