dirdiff: Filter line endings when comparing file contents (bgo#547743)
This commit changes the behaviour of folder comparisons for text files
where the only difference is the end-of-line characters. For symmetry
with file comparisons, folder comparisons now treat these as being
"Same, with filtering" files, indicated by italics in the file name.
Along with the new notifications in file comparisons, this should get
rid of a lot of the confusion users had when folder comparison told
them that files were different, but file comparison said they were
identical.

With this change, we now *always* need to read in the whole file (if
we think it's not binary) because we're effectively always applying a
text filter.

This commit also adds the final piece of handling for the new
'apply-text-filters' preference, which was introduced precisely because
users may want to opt out of the new, slow always-read-all-the-files
behaviour.