meld.vc: Fix caching API to disallow inadventantly discarding data
The caching API that VCs needed to implement suggested that it would be
fine to try and cache a subdirectory independently. This never worked.

Now you can only call cache without args, and it will (shoudl) always
cache the location under the VC. Updating the cache is left for other
methods.