dirdiff, filediff: Add a hack to revert overlay scrolling (bgo#748219)
In some cases, overlay scrolling appears to mess with the ability to
interact with, for example, Meld's merge buttons. As such, this patch
reverts to using old-style scrollbars for folder and file comparisons,
since they're the ones likely to be affected.

It might actually be quite nice for Meld to enable these, if this gets
fixed or we can figure out a different workaround.