filediff: Add hack to work around double setting of filechooser paths
This causes a crash on at least some versions of GKT+, but we kind of
need it (for now at least) to handle --output setting. We should be
able to work around this better, but for now this hack will do the
job.

Also, this bug has already been fixed in GTK+, but only in 3.18 so
it's well above what we're going to require for this release of Meld.