Move filediff to using GtkSourceFileSaver
This patch is a first pass at using the saver approach, and includes
lots of async-related bits and pieces.

Getting this right is a real challenge, since we have plenty of error
handling code that assumes that it can call save and know whether the
file has saved when that returns.