filediff: Use locale-based default encodings (bgo#752307, bgo#753427)
Taking some inspiration from GtkSourceViewLoader, this patch changes
our list of encodings-to-try when loading files to follow a set
pattern.

While this is a behaviour change, it's basically just adding the
locale-specific charset and UTF-16 to the list to be loaded; it's
unlikely that the latin-1 -> iso-8859-15 is going to do anything but
load more files correctly.

The default list of encodings in our schema is left as ["utf8"], so
that if users add things to this list, there is at least a reasonable
hint that they should add then *after* UTF-8 to preserve sane loading.