meldapp: Monkey-patch optparse's gettext alias for ugettext compat
Meld uses and assumes ugettext for our translations, and this doesn't
interact well with optparse's use of gettext, when it tries to mash
together its bytestring translations with our unicode translations.

While this is an ugly hack, the only other simple option is not using
optparse, which is a longer-term goal but much too invasive for a
quick fix.