filediff: Offer to open binary files externally (bgo#665154)
If we think we've got a binary file, currently Meld just refuses to
display it. It's somewhat more useful to to offer to open this file in
whatever program we think will be able to view it. This is primarily
for file comparisons started from folder comparisons, though VC
comparisons will also benefit somewhat.