meldapp, meldwindow: Rework window deletion logic for async closing
Because we're not doing asynchronous saving, we need a significantly
more complicated flow of should-I-close logic. Currently this works by
 having MeldWindow register whether it thinks it should try to close,
and having it destroy itself if it thinks it's trying to close and
doesn't have any tabs left. MeldApp now does something similar, and
dies if its last window ever destroys itself.

The delete/destroy logic in MeldWindow.page_removed still seems weird
to me, but it's definitely necessary for the window to close correctly
when e.g., an unsaved file comparison is present, the user saves it
from the quit dialog, and the async save finishes.