vc.git: Support showing (partially) staged status of files (bgo#755112)
The use case here is that it's nice to know what you still need to look
at, particularly in more complicated merges. We only differentiate
between files that are unstaged, fully staged and partially staged.

Actually having a real staging area view is an orthogonal consideration,
and would still be nice.