sourceview: Base x-coords for drawing on clip area, not allocation
The visible effect of this bug was that when scrolling horizontally,
areas would sometimes be left unpainted. This was because of the new
text view pixel cache, which sets clip areas to exceed the actual
visible area.

We've already fixed this for vertical allocations; this commit just
fixes the bug for horizontal coordinates as well.