filediff: Use the MeldBuffer-associated source file to get metadata set
Because we were recreating the source file that we passed along to the
loader, none of the interesting attributes that the loader sets (e.g.,
the newline type or encoding) ended up set on the file we actually had
associated with our MeldBuffer. Since we'd like to move to using the
loaded sourcefile for whatever attributes we can, we now use this when
loading.