Dim text ignored by filters
Filtered text that is not indicated somehow can lead to confusion,
because for example Melds comparison algorithm does not distinguish
between a blank line and a line that matches a text filter.

Now the filtered text is dimmed, that means the foreground color of
filtered text is changed to light grey to indicate that it is ignored by
Melds comparison algorithm.

https://bugzilla.gnome.org/show_bug.cgi?id=633515