Remove tab list from menu, replacing with keybinding actions
This commit creates a new MeldNotebook class for handling some contained
notebook behaviours and keybindings. Since we're looking to reduce the
overall UI handled by UIManager, and because it's no longer standard to
keep a list of open tabs as menu items, we're dropping that bit of the
UI and accompanying handling. This reimplements only the Alt+number
keybindings for quick tab switching.