ui.historyentry: Don't use SafeConfigParser's interpolation (bgo#757659)
SafeConfigParser and ConfigParser both support magical interpolation
syntax that we don't use. Better yet, these classes will happily write
out values that they can't read back in, and traceback when you try.

Specifically, in this case a commit message containing "%PARAM1%" broke
the parsers because they considered this to be an invalid interpolation
string, but only errored the *next* time you tried to change the config
file.

This commit just moves us to using RawConfigParser because we don't
want the magical behaviour.