New text-filter mechanism to prevent interference of multiple filters
By now, each text filter cuts out the out-filtered parts, so all following
filters operate on the already modified text. This leads to mistakes.

The new mechanism first stores, which parts have to be cut out for
each filter, merges these ranges and in the end actualy cuts the text.

https://bugzilla.gnome.org/show_bug.cgi?id=758478