meldwindow: Ensure UIManager update, and stop updates (bgo#755407)
When closing a window, if there are any pending UIManager updates, we
get lots of extremely ugly, important-sounding Gtk-CRITICAL logging on
the console.

This change just ensures that any UIManager updates caused by the tab
removal are done before we move on, and stops doing any more updates if
we have no more tabs. Since we always close all tabs when closing a
window, this should stop the problem from occurring.