data/meld.css: Make the LinkMap stand out from the background
Some users have very reasonably complained that having the LinkMap have
the same colour as the GtkSourceView makes it hard to see what kind of
indentation level you're dealing with on right-hand panes. While we'd
need to add backgrounds to the GtkSourceGutters as well to actually fix
this, giving the Linkmap a background helps a little.