Bugfix: error dialog for multiple line text filters fixed
In patch eee58a571b1dc75f0d85d266aecec48a52df38c3 we applied
a new text filter mechanism. This broke the error message when using
multiple line text filters.

Because the filters are not applied one by one anymore, the error dialog
does not know which filter is wrong.

With this patch the filter is no longer mentioned in the error message.

https://bugzilla.gnome.org/show_bug.cgi?id=761028