meld.vc: Add a REMOVED/MISSING cache for speed issues (bgo#761789)
In large repositories, we were checking the whole list of file states
for every directory to identify files in that directory that weren't
present on the filesystem.

With this change, each VC now maintains a cache of files with REMOVED
and MISSING states, keyed by their parent directory, reducing the
expected complexity of inserting these files into the VC model.