misc: Try to auto-detect whether to use our dark variant (bgo#764713)
Unfortunately, there's no actual way to know whether a GTK+ theme is
a dark theme variant. In theory we could come up with better
heuristics, but in reality we'll just trust the prefer-dark-theme
setting.

For completeness (mostly for testing) we also check to see whether a
dark variant has been requested via the environment.

These checks vaguely mimic what GTK+ inspector currently does for
theme handling, so this is probably the best we can do for now.