filediff, meldbuffer: Support file comparisons from pipes (bgo#765021)
With the move to GtkSourceFileLoader, etc., we unintentionally lost the
ability to read from named pipes, because yours truly had no idea this
worked at all. This patch is a slightly hacky approach to handle
switching to the stream-based loading in GtkSourceFileLoader if it
appears that the GFile we have doesn't point to a regular file.