filediff: Fix DiffChunk sorting tuple-of-maybe-tuples (bgo#767744)
We were sorting a tuple of DiffChunk-or-None elements, which no longer
compare in Python 3. In addition, the sort (which I believe exists only
so that we rehighlight in a predictable order) was doing a native
sort, which resulted in sorting first based on chunk type.

With this change, we handle None in merged chunks correctly (by
ignoring it) and sort based on a chunk's start location in the middle
column.