filediff, diffutil: Scroll all chunks on screen at once (bgo#767981)
When we're navigating by chunk, our previous code found the correct
chunk in the focused pane, placed the cursor on the first line and
scrolled to the cursor. This works in most cases, because our scroll
alignment handling takes care of the rest. However, for chunks at the
start or end of the file (especially for very unequal file lengths)
this means you can end up not actually scrolling the real change on
screen, for example if you're on the empty side of an insertion chunk.

This change does a slightly inefficient thing and actually gets the
offsets for all panes of an aligned chunk and scrolls the first line of
each pane's chunk on screen. We then re-scroll to the cursor, just in
case our scroll alignment handling has changed the scroll too much; it
shouldn't, but if it does it would be *really* irritating.